requirements.txt
appdirs==1.4.0
ExifRead==2.1.2
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
six==1.10.0
